homework4/requirements.txt
###### Requirements without Version Specifiers ######
psutil
PyYAML

###### Requirements with Version Specifiers ######
python_version == '2.7'